+++ tests/botan.rs
@@ -1,10 +1,10 @@
 #![cfg(feature = "x509-parser")]
 
-use rcgen::DnValue;
 use rcgen::{BasicConstraints, Certificate, CertificateParams, DnType, IsCa};
 use rcgen::{
 	CertificateRevocationList, CertificateRevocationListParams, RevocationReason, RevokedCertParams,
 };
+use rcgen::{DnValue, KeyPair};
 use rcgen::{KeyUsagePurpose, SerialNumber};
 use time::{Duration, OffsetDateTime};
 
@@ -172,7 +172,6 @@ fn test_botan_separate_ca() {
 #[cfg(feature = "x509-parser")]
 #[test]
 fn test_botan_imported_ca() {
-	use std::convert::TryInto;
 	let mut params = default_params();
 	params.is_ca = IsCa::Ca(BasicConstraints::Unconstrained);
 	let ca_cert = Certificate::from_params(params).unwrap();
@@ -182,7 +181,7 @@ fn test_botan_imported_ca() {
 		ca_cert.serialize_private_key_der(),
 	);
 
-	let ca_key_pair = ca_key_der.as_slice().try_into().unwrap();
+	let ca_key_pair = KeyPair::from_der(ca_key_der.as_slice()).unwrap();
 	let imported_ca_cert_params =
 		CertificateParams::from_ca_cert_der(ca_cert_der.as_slice(), ca_key_pair).unwrap();
 	let imported_ca_cert = Certificate::from_params(imported_ca_cert_params).unwrap();
@@ -205,7 +204,6 @@ fn test_botan_imported_ca() {
 #[cfg(feature = "x509-parser")]
 #[test]
 fn test_botan_imported_ca_with_printable_string() {
-	use std::convert::TryInto;
 	let mut params = default_params();
 	params.distinguished_name.push(
 		DnType::CountryName,
@@ -219,7 +217,7 @@ fn test_botan_imported_ca_with_printable_string() {
 		ca_cert.serialize_private_key_der(),
 	);
 
-	let ca_key_pair = ca_key_der.as_slice().try_into().unwrap();
+	let ca_key_pair = KeyPair::from_der(ca_key_der.as_slice()).unwrap();
 	let imported_ca_cert_params =
 		CertificateParams::from_ca_cert_der(ca_cert_der.as_slice(), ca_key_pair).unwrap();
 	let imported_ca_cert = Certificate::from_params(imported_ca_cert_params).unwrap();